+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 ## 0.0.0
 * Add `do_load` 
 * Enable licensing 
+* Make scripts/styles consistent